# part3/hbnb/tests/Schemathesis/swagger.json
{
    "swagger": "2.0",
    "basePath": "/",
    "paths": {
...
        "/api/v1/users/": {
            "post": {
                "responses": {
                    "400": {
                        "description": "Email already registered or invalid input"
                    },
                    "201": {
                        "description": "User successfully created"
                    }
                },
                "summary": "Register new user",
                "operationId": "post_user_list",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "tags": [
                    "auth"
                ]
            },
            "get": {
                "responses": {
                    "404": {
                        "description": "User not found"
                    },
                    "200": {
                        "description": "User(s) found"
                    }
                },
                "summary": "Get all users or by email query",
                "operationId": "get_user_list",
                "parameters": [
                    {
                        "description": "Filter user by email (optional)",
                        "name": "email",
                        "type": "string",
                        "in": "query"
                    }
                ],
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/v1/users/admin_creation": {
            "post": {
                "responses": {
                    "401": {
                        "description": "You are not allowed to create an admin"
                    },
                    "400": {
                        "description": "Email already registered or invalid input"
                    },
                    "201": {
                        "description": "Admin successfully created"
                    }
                },
                "operationId": "post_admin_creation",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "tags": [
                    "auth"
                ]
            }
        },
        "/api/v1/users/login": {
            "post": {
                "responses": {
                    "404": {
                        "description": "User not found"
                    },
                    "401": {
                        "description": "Bad credentials"
                    },
                    "400": {
                        "description": "Pydantic validation error"
                    },
                    "201": {
                        "description": "Token created"
                    }
                },
                "summary": "Login the user",
                "operationId": "post_login",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Login"
                        }
                    }
                ],
                "tags": [
                    "auth"
                ]
            }
        },
...
        "/api/v1/users/{user_id}": {
            "parameters": [
                {
                    "name": "user_id",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "400": {
                        "description": "Invalid UUID format"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "200": {
                        "description": "user details retrieved successfully"
                    }
                },
                "summary": "Get user details by ID",
                "operationId": "get_user_resource",
                "tags": [
                    "auth"
                ]
            },
            "put": {
                "responses": {
                    "404": {
                        "description": "User not found"
                    },
                    "403": {
                        "description": "Permission error"
                    },
                    "400": {
                        "description": "Invalid input or UUID format"
                    },
                    "200": {
                        "description": "User successfully updated"
                    }
                },
                "summary": "Update an existing user",
                "operationId": "put_user_resource",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UserUpdate"
                        }
                    }
                ],
                "tags": [
                    "auth"
                ]
            },
            "delete": {
                "responses": {
                    "404": {
                        "description": "User not found"
                    },
                    "403": {
                        "description": "Permission error"
                    },
                    "400": {
                        "description": "Invalide UUID format"
                    },
                    "200": {
                        "description": "User deleted successfully"
                    }
                },
                "operationId": "delete_user_resource",
                "tags": [
                    "auth"
                ]
            }
        }
    },
    "info": {
        "title": "HBnB API",
        "version": "1.0",
        "description": "HBnB Application API"
    },
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
    "securityDefinitions": {
        "Bearer Auth": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization",
            "description": "Enter 'Bearer' + your token here"
        }
    },
    "security": [
        {
            "Bearer Auth": []
        }
    ],
...
    "definitions": {
        "User": {
            "required": [
                "email",
                "first_name",
                "last_name",
                "password"
            ],
            "properties": {
                "first_name": {
                    "type": "string",
                    "description": "First name of user"
                },
                "last_name": {
                    "type": "string",
                    "description": "Last name of user"
                },
                "email": {
                    "type": "string",
                    "description": "Email of user"
                },
                "password": {
                    "type": "string",
                    "description": "Password of user"
                },
                "photo_url": {
                    "type": "string",
                    "description": "Avatar of user"
                }
            },
            "type": "object"
        },
        "UserUpdate": {
            "properties": {
                "first_name": {
                    "type": "string",
                    "description": "First name of user"
                },
                "last_name": {
                    "type": "string",
                    "description": "Last name of user"
                },
                "email": {
                    "type": "string",
                    "description": "Email of user"
                },
                "password": {
                    "type": "string",
                    "description": "Password of user"
                },
                "photo_url": {
                    "type": "string",
                    "description": "Avatar of user"
                }
            },
            "type": "object"
        },
        "Login": {
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "description": "User email"
                },
                "password": {
                    "type": "string",
                    "description": "User password"
                }
            },
            "type": "object"
        },
...
    },
...
}